organize factory files
