+++ lib/user/store/elasticsearch.js
@@ -1,4 +1,4 @@
-import UserBase from './userBase.js'
+import UserBase from './base.js'
 
 class UserRepoElasticsearch extends UserBase {
   async authenticate(_authTry) {
