+++ lib/user/store/mongodb.js
@@ -1,4 +1,4 @@
-import UserBase from './userBase.js'
+import UserBase from './base.js'
 
 class UserRepoMongoDB extends UserBase {
   async authenticate(_authTry) {
