+++ lib/user/store/mysql.js
@@ -1,8 +1,8 @@
-import Mysql from '../mysql.js'
-import Config from '../config.js'
-import UserBase from './userBase.js'
-import Permission from '../permission.js'
-import { mapToDbColumn } from '../util.js'
+import Mysql from '../../mysql.js'
+import Config from '../../config.js'
+import UserBase from './base.js'
+import Permission from '../../permission.js'
+import { mapToDbColumn } from '../../util.js'
 
 const userDbMap = { id: 'nt_user_id', gid: 'nt_group_id' }
 const boolFields = ['is_admin', 'deleted']
