+++ lib/user/store/toml.js
@@ -4,8 +4,8 @@ import { fileURLToPath } from 'node:url'
 
 import { parse, stringify } from 'smol-toml'
 
-import Config from './config.js'
-import UserBase from './userBase.js'
+import Config from '../../config.js'
+import UserBase from './base.js'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
 const boolFields = ['is_admin', 'deleted']
@@ -13,7 +13,7 @@ const boolFields = ['is_admin', 'deleted']
 function resolveStorePath(filename) {
   const base = process.env.NICTOOL_DATA_STORE_PATH
   if (base) return path.join(base, filename)
-  return path.resolve(__dirname, '../conf.d', filename)
+  return path.resolve(__dirname, '../../../conf.d', filename)
 }
 
 class UserRepoTOML extends UserBase {
