+++ lib/zone/index.js
@@ -3,16 +3,16 @@ const storeType = process.env.NICTOOL_DATA_STORE ?? 'mysql'
 let RepoClass
 switch (storeType) {
   case 'toml':
-    RepoClass = (await import('./toml.js')).default
+    RepoClass = (await import('./store/toml.js')).default
     break
   case 'mongodb':
-    RepoClass = (await import('./mongodb.js')).default
+    RepoClass = (await import('./store/mongodb.js')).default
     break
   case 'elasticsearch':
-    RepoClass = (await import('./elasticsearch.js')).default
+    RepoClass = (await import('./store/elasticsearch.js')).default
     break
   default:
-    RepoClass = (await import('./mysql.js')).default
+    RepoClass = (await import('./store/mysql.js')).default
 }
 
 export default new RepoClass()
