+++ lib/zone/store/elasticsearch.js
@@ -1,4 +1,4 @@
-import ZoneBase from './zoneBase.js'
+import ZoneBase from './base.js'
 
 class ZoneRepoElasticsearch extends ZoneBase {
   async get(_args) {
