+++ lib/zone/store/mongodb.js
@@ -1,4 +1,4 @@
-import ZoneBase from './zoneBase.js'
+import ZoneBase from './base.js'
 
 class ZoneRepoMongoDB extends ZoneBase {
   async get(_args) {
