+++ lib/zone/store/mysql.js
@@ -1,6 +1,6 @@
-import Mysql from '../mysql.js'
-import ZoneBase from './zoneBase.js'
-import { mapToDbColumn } from '../util.js'
+import Mysql from '../../mysql.js'
+import ZoneBase from './base.js'
+import { mapToDbColumn } from '../../util.js'
 
 const zoneDbMap = { id: 'nt_zone_id', gid: 'nt_group_id' }
 const boolFields = ['deleted']
