+++ lib/zone/store/toml.js
@@ -4,7 +4,7 @@ import { fileURLToPath } from 'node:url'
 
 import { parse, stringify } from 'smol-toml'
 
-import ZoneBase from './zoneBase.js'
+import ZoneBase from './base.js'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
@@ -13,7 +13,7 @@ const zoneDefaults = { minimum: 3600, ttl: 3600, refresh: 86400, retry: 7200, ex
 function resolveStorePath(filename) {
   const base = process.env.NICTOOL_DATA_STORE_PATH
   if (base) return path.join(base, filename)
-  return path.resolve(__dirname, '../../conf.d', filename)
+  return path.resolve(__dirname, '../../../conf.d', filename)
 }
 
 class ZoneRepoTOML extends ZoneBase {
