+++ lib/zone_record/index.js
@@ -3,10 +3,10 @@ const storeType = process.env.NICTOOL_DATA_STORE ?? 'mysql'
 let RepoClass
 switch (storeType) {
   case 'toml':
-    RepoClass = (await import('./toml.js')).default
+    RepoClass = (await import('./store/toml.js')).default
     break
   default:
-    RepoClass = (await import('./mysql.js')).default
+    RepoClass = (await import('./store/mysql.js')).default
 }
 
 export default new RepoClass()
