+++ lib/zone_record/store/mysql.js
@@ -1,7 +1,7 @@
 import * as RR from '@nictool/dns-resource-record'
 
-import Mysql from '../mysql.js'
-import { mapToDbColumn } from '../util.js'
+import Mysql from '../../mysql.js'
+import { mapToDbColumn } from '../../util.js'
 
 const zrDbMap = { id: 'nt_zone_record_id', zid: 'nt_zone_id', owner: 'name' }
 const boolFields = ['deleted']
