+++ lib/zone_record/store/toml.js
@@ -9,7 +9,7 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url))
 function resolveStorePath(filename) {
   const base = process.env.NICTOOL_DATA_STORE_PATH
   if (base) return path.join(base, filename)
-  return path.resolve(__dirname, '../../conf.d', filename)
+  return path.resolve(__dirname, '../../../conf.d', filename)
 }
 
 class ZoneRecordRepoTOML {
